requirements.txt
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyGithub==1.59.1
python-dotenv==1.0.0

# Semantic Search Dependencies
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.6.1
tqdm==4.66.1
jsonlines==4.0.0
python-slugify==8.0.1
